FSPT-438: moves status integration test to unit test
Now that we have collection metadata factories we can stub out the
longer scenario as a unit test to avoid any back and forth with the
database.
